requirements.txt
boto3==1.24.28
PyYAML==6.0.1
Requests==2.31.0
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1
scikit-learn==1.4.2
pytest==8.2.0